chore: refactor export, make it work with Excel and CSV
