+++ desktop/src/scenes/Export/export.jsx
@@ -1,8 +1,8 @@
-import React, { Component } from 'react'
+import { Component } from 'react'
 import PropTypes from 'prop-types'
 
 import { withStyles } from '@material-ui/core/styles'
-import ExportContainerCSV from '~/containers/Export/exportContainerCSV'
+import ExportContainer from '~/containers/Export/exportContainer'
 
 const styles = () => ({
   hero: {
@@ -16,11 +16,14 @@ const styles = () => ({
 
 class Export extends Component {
   render() {
-    const { classes } = this.props
+    const { classes, history } = this.props
+
     return (
       <div className={classes.hero}>
-        {/* this is where export container type is atm, id keep this structure for now until we know if we want to have both export files */}
-        <ExportContainerCSV />
+        <ExportContainer
+          onSubmit={() => history.push(`/`)}
+          onCancel={() => window.history.back()}
+        />
       </div>
     )
   }
@@ -29,4 +32,5 @@ class Export extends Component {
 Export.propTypes = {
   classes: PropTypes.object.isRequired,
 }
+
 export default withStyles(styles)(Export)
